Funciona la opción de pagar
Se descuenta el saldo del cliente pero cuando se guarda en la base de
datos se redondea. Además en el label de precio no se muestra con dos
decimales.

+++ src/controllers/cliente/CarnetController.java
@@ -33,13 +33,13 @@ public class CarnetController implements Initializable {
     private PasswordField txtContrasenia;
 
     private Cliente cliente;
-    private int cantPagar;
+    private double cantPagar;
     
     public void setCliente(Cliente cliente){
         this.cliente = cliente;
     }
     
-    public void setCantPagar(int cantidad){
+    public void setCantPagar(double cantidad){
         this.cantPagar = cantidad;
     }
